change to use modules - sort array on display - ad  : and space to list item
