+++ src/index.html
@@ -19,9 +19,6 @@
           </div>
           <div>
             <ul class="scoreList" id="gameScoreList">
-              <li class="scoreItem">Name:100</li>
-              <li class="scoreItem greyItem">Name:20</li>
-              <li class="scoreItem">Name:42</li>
             </ul>
           </div>
         </article>
